refactor: 계층 간 전달하는 인자 컨벤션 개선
