+++ src/main/java/roomescape/reservation/application/ReservationService.java
@@ -1,10 +1,8 @@
 package roomescape.reservation.application;
 
-import roomescape.reservation.domain.ReservationId;
 import roomescape.reservation.ui.dto.AvailableReservationTimeWebResponse;
 import roomescape.reservation.ui.dto.CreateReservationWebRequest;
 import roomescape.reservation.ui.dto.ReservationResponse;
-import roomescape.theme.domain.ThemeId;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -13,9 +11,9 @@ public interface ReservationService {
 
     List<ReservationResponse> getAll();
 
-    List<AvailableReservationTimeWebResponse> getAvailable(LocalDate date, ThemeId themeId);
+    List<AvailableReservationTimeWebResponse> getAvailable(LocalDate date, Long id);
 
     ReservationResponse create(CreateReservationWebRequest createReservationWebRequest);
 
-    void delete(ReservationId id);
+    void delete(Long id);
 }
