+++ src/main/java/roomescape/reservation/application/usecase/ReservationQueryUseCase.java
@@ -2,6 +2,7 @@ package roomescape.reservation.application.usecase;
 
 import roomescape.reservation.application.dto.AvailableReservationTimeServiceRequest;
 import roomescape.reservation.application.dto.AvailableReservationTimeServiceResponse;
+import roomescape.reservation.application.dto.ThemeToBookCountServiceResponse;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.theme.domain.ThemeId;
